Organize the project in packages
