+++ psrm/enums/aktivitettype.py
