+++ psrm/enums/artkode.py
