+++ psrm/enums/typekategori.py
