+++ psrm/enums/typekode.py
